Moved COW implementation to its own class
